Release GVL while RSA key generation
Define ossl_generate_cb_stop as a unblock function for
rb_thread_blocking_region. The unblock function sets a stop bit, and
the callback function prepared for OpenSSL tries to stop the BN
generate iteration.

The callback function prepared for OpenSSL is also used for callback in
Ruby world, so we cannot release GVL when a caller supplies callback
proc. For example, it relases GVL for following code;

  OpenSSL::PKey::RSA.new(2048)

but does not release for following code;

  OpenSSL::PKey::RSA.new(2048) do
    print '.'
  end